requirements.txt
sanic==20.12.2
environs==9.3.1
sanic-envconfig==1.3.4
secure==0.2.1
sqlalchemy-migrate==0.13.0
Sanic-Auth==0.3.0
databases[postgresql]
psycopg2==2.8.6
bcrypt==3.2.0
pytest-sanic==1.6.2
